[trainer] refactor: fold sample-level dispatch into ReplayBufferAsync
Make it a first-class sampler config (dispatch_mode=sample_level) instead
of a separate plugin file under v1.
