+++ tests/trainer/ppo/v1/test_sample_level_sampler_on_cpu.py
@@ -20,8 +20,8 @@ import pytest
 import torch
 import transfer_queue as tq
 
-from verl.trainer.ppo.v1.sample_level_sampler import (
-    SampleLevelReplayBuffer,
+from verl.trainer.ppo.v1.replay_buffer import (
+    ReplayBufferAsync,
     compute_sample_level_dispatch,
     count_inflight_samples,
 )
@@ -143,13 +143,16 @@ def test_prompt_level_would_wait_sample_level_does_not(tq_init, partition_id):
         calls.append(n)
         return n
 
-    rb = SampleLevelReplayBuffer(
+    rb = ReplayBufferAsync(
         trainer_mode="colocate_async",
         trainer_config={},
         max_off_policy_threshold=8,
         max_off_policy_strategy="drop",
-        sampler_kwargs={"group_size": 2, "max_inflight_samples": 6},
+        sampler_kwargs={},
         refill_fn=refill_fn,
+        dispatch_mode="sample_level",
+        group_size=2,
+        max_inflight_samples=6,
     )
     try:
         rb._sync_metadata_from_transfer_queue()
@@ -163,6 +166,49 @@ def test_prompt_level_would_wait_sample_level_does_not(tq_init, partition_id):
             tq.kv_clear(keys=keys, partition_id=partition_id)
 
 
+def test_build_replay_buffer_wires_sample_level_from_rollout_n():
+    from omegaconf import OmegaConf
+
+    from verl.trainer.ppo.v1.trainer_base import PPOTrainer
+
+    class _T(PPOTrainer):
+        def on_step_end(self):
+            pass
+
+        def on_sample_end(self):
+            pass
+
+    trainer = _T.__new__(_T)
+    trainer.trainer_mode = "colocate_async"
+    trainer.config = OmegaConf.create(
+        {
+            "algorithm": {"filter_groups": {"enable": False}},
+            "data": {"train_batch_size": 8},
+            "actor_rollout_ref": {"rollout": {"n": 8}},
+            "reward": {"reward_model": {"enable": False, "enable_resource_pool": False}},
+            "trainer": {
+                "v1": {
+                    "colocate_async": {},
+                    "sampler": {
+                        "custom_sampler": None,
+                        "max_off_policy_threshold": 8,
+                        "max_off_policy_strategy": "drop",
+                        "sampler_kwargs": {},
+                        "dispatch_mode": "sample_level",
+                    },
+                }
+            },
+        }
+    )
+    trainer._add_prompts_to_generate = lambda n: n
+    rb = trainer._build_replay_buffer()
+    assert isinstance(rb, ReplayBufferAsync)
+    assert rb.dispatch_mode == "sample_level"
+    assert rb.group_size == 8
+    assert rb.max_inflight_samples == 64
+    assert rb.trainer_owns_dispatch is False
+
+
 def test_trainer_skips_batch_dump_when_sampler_owns_dispatch():
     from verl.trainer.ppo.v1.trainer_base import PPOTrainer
 
@@ -174,8 +220,8 @@ def test_trainer_skips_batch_dump_when_sampler_owns_dispatch():
             pass
 
     trainer = _T.__new__(_T)
-    trainer.replay_buffer = SampleLevelReplayBuffer.__new__(SampleLevelReplayBuffer)
-    trainer.replay_buffer.trainer_owns_dispatch = False
+    trainer.replay_buffer = ReplayBufferAsync.__new__(ReplayBufferAsync)
+    trainer.replay_buffer.dispatch_mode = "sample_level"
     trainer._add_batch_to_generate = MagicMock()
 
     assert trainer._should_add_batch_to_generate() is False
