+++ verl/trainer/config/_generated_ppo_megatron_trainer.yaml
@@ -1034,6 +1034,9 @@ trainer:
       max_off_policy_threshold: 8
       max_off_policy_strategy: drop
       sync_refill_failed_groups: false
+      dispatch_mode: batch
+      group_size: null
+      max_inflight_samples: null
       custom_sampler:
         path: null
         name: null
