+++ verl/trainer/config/_generated_ppo_torchtitan_trainer.yaml
@@ -929,6 +929,9 @@ trainer:
       max_off_policy_threshold: 8
       max_off_policy_strategy: drop
       sync_refill_failed_groups: false
+      dispatch_mode: batch
+      group_size: null
+      max_inflight_samples: null
       custom_sampler:
         path: null
         name: null
